モジュール作成：入力関連
