+++ src/component/App.jsx
@@ -1,23 +1,11 @@
 import React from 'react'
 import Rect from './Rect.jsx'
 import Handleupload from './Handleupload.jsx'
+import Input from './Input.jsx'
 
 
-const TextInput = (props) => {
-  return <input ref={props.inputRef} />
-}
 
 export default class App extends React.Component {
-
-    // フォーカス
-    focusInput = () => {
-      this.input.focus()
-    }
-
-    // 入力連動
-    changeInput = () => {
-      this.textInput.value = this.input.value
-    }
 
     render () {
       return (
@@ -29,16 +17,8 @@ export default class App extends React.Component {
           <Rect num={3} bgcolor='#2020e0' />
           <h1>File Upload</h1>
           <Handleupload />
-          <div>
-            <div>
-              {/* <input type="text" ref={(input) => { this.input = input }} /> */}
-              <button onClick={this.focusInput}>入力フォーカス</button>
-            </div>
-          </div>
-          <div>
-            <input type="text" ref={(input) => { this.input = input }} onChange={this.changeInput} />
-            <TextInput inputRef={el => this.textInput = el} />
-          </div>
+          <h1>Text Input</h1>
+          <Input />
         </>
       )
     }
